+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+
+## [0.4.2] - 2024-04-10
 ### Fixed
 - count() now defaults to '*' only if no limit or offset is set
 - Repository->fetchChunked() is now typehinted too
 - InsertRepositoryEvent now correctly sets Query on the reused entity.
 - PHPStan lvl8
-
 
 ## [0.4.1] - 2024-03-19
 ### Fixed
@@ -50,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial version
 
-[Unreleased]: https://github.com/efabrica-team/nette-database-repository/compare/0.4.1...main
+[Unreleased]: https://github.com/efabrica-team/nette-database-repository/compare/0.4.2...main
+[0.4.2]: https://github.com/efabrica-team/nette-database-repository/compare/0.4.1...0.4.2
 [0.4.1]: https://github.com/efabrica-team/nette-database-repository/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/efabrica-team/nette-database-repository/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/efabrica-team/nette-database-repository/compare/0.2.1...0.3.0